Add `home` view and `teacher default` view
Add end-to-end tests and nest describe blocks inside the primary
describe block to separate end-to-end tests for each component.

Add some default routing to the application.

Add a `Home` component to the application for default unauthenticated
view.

Add a `TeacherHome` component to the apaplication for default
authenticated teacher view. This is the initial thing teachers will see
just after logging in.

Add some teacher-based routing.

Add font-awesome fonts.

Update borders on `header-nav` element (based on a suggestion) to
exclude the white border on its left side.